project restructure
