+++ Sources/braze/src/androidTest/AndroidManifest.xml
