+++ Sources/braze/src/main/AndroidManifest.xml
