+++ Sources/braze/src/main/res/xml/network_security_config.xml
