+++ Sources/build.gradle
@@ -4,9 +4,10 @@ buildscript {
     repositories {
         google()
         jcenter()
+        
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.4.1'
+        classpath 'com.android.tools.build:gradle:3.5.0'
         
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
