+++ TealiumBrazeExample/settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':braze'
+include ':app'
